Refactor input so as to separate the mongodb specific code
