+++ converter/commands/convert.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/activecm/ipfix-rita/converter/environment"
 	"github.com/activecm/ipfix-rita/converter/filter"
-	input "github.com/activecm/ipfix-rita/converter/input/mgologstash"
+	input "github.com/activecm/ipfix-rita/converter/input/logstash/mongodb"
 	"github.com/activecm/ipfix-rita/converter/logging"
 	"github.com/activecm/ipfix-rita/converter/output"
 	batchRITAOutput "github.com/activecm/ipfix-rita/converter/output/rita/batch/dates"
