+++ converter/input/logstash/mongodb/database.go
@@ -1,12 +1,12 @@
-package mgologstash
+package mongodb
 
 import (
 	"time"
 
 	"github.com/activecm/ipfix-rita/converter/config"
 	"github.com/activecm/ipfix-rita/converter/database"
-	"github.com/pkg/errors"
 	mgo "github.com/globalsign/mgo"
+	"github.com/pkg/errors"
 )
 
 //LogstashMongoInputDB wraps a *mgo.Database connected
